Move concepts logic into concepts doc viewer
